+++ RELEASE_NOTES.md
@@ -4,6 +4,9 @@ Changes:
 
   * Added partition number and library alias to Cell Ranger sample sheets
   * Renamed library dilutions to library aliquots
+  * Renamed sequencing order states
+    * Active -> Outstanding
+    * Pending -> In-Progress
   * Fixed missing storage location for library aliquots in Pinery-MISO
 
 Upgrade Notes: